[command] move the code to vm1
