+++ cmd/starcoin/src/chain/mod.rs
@@ -9,7 +9,7 @@ mod get_txn_cmd;
 mod get_txn_info_cmd;
 mod get_txn_info_list_cmd;
 mod get_txn_infos_cmd;
-pub mod get_txn_proof_cmd;
+mod get_txn_proof_cmd;
 mod info_cmd;
 mod list_block_cmd;
 
@@ -21,5 +21,6 @@ pub use get_txn_cmd::*;
 pub use get_txn_info_cmd::*;
 pub use get_txn_info_list_cmd::*;
 pub use get_txn_infos_cmd::*;
+pub use get_txn_proof_cmd::*;
 pub use info_cmd::*;
 pub use list_block_cmd::*;
