+++ cmd/starcoin/src/lib.rs
@@ -100,13 +100,13 @@ pub fn add_command(
                 .subcommand(chain::InfoCommand)
                 .subcommand(chain::GetBlockCommand)
                 .subcommand(chain::ListBlockCommand)
-                .subcommand(chain2::GetTransactionCommand)
-                .subcommand(chain2::GetTxnInfosCommand)
-                .subcommand(chain2::GetTransactionInfoCommand)
-                .subcommand(chain2::GetEventsCommand)
+                .subcommand(chain::GetTransactionCommand)
+                .subcommand(chain::GetTxnInfosCommand)
+                .subcommand(chain::GetTransactionInfoCommand)
+                .subcommand(chain::GetEventsCommand)
                 .subcommand(chain::EpochInfoCommand)
-                .subcommand(chain2::GetTransactionInfoListCommand)
-                .subcommand(chain2::get_txn_proof_cmd::GetTransactionProofCommand)
+                .subcommand(chain::GetTransactionInfoListCommand)
+                .subcommand(chain::GetTransactionProofCommand)
                 .subcommand(chain::GetBlockInfoCommand),
         )
         .command(
