--- cmd/starcoin/src/subcommand_vm2/chain/get_txn_infos_cmd.rs
@@ -1,38 +1,0 @@
-// Copyright (c) The Starcoin Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-use crate::cli_state::CliState;
-use crate::StarcoinOpt;
-use anyhow::Result;
-use clap::Parser;
-use scmd::{CommandAction, ExecContext};
-use starcoin_crypto::HashValue;
-use starcoin_vm2_types::view::TransactionInfoView;
-
-/// Get transaction infos by block hash.
-#[derive(Debug, Parser)]
-#[clap(name = "get-txn-infos")]
-pub struct GetTxnInfosOpt {
-    #[clap(name = "block-hash")]
-    block_hash: HashValue,
-}
-
-pub struct GetTxnInfosCommand;
-
-impl CommandAction for GetTxnInfosCommand {
-    type State = CliState;
-    type GlobalOpt = StarcoinOpt;
-    type Opt = GetTxnInfosOpt;
-    type ReturnItem = Vec<TransactionInfoView>;
-
-    fn run(
-        &self,
-        ctx: &ExecContext<Self::State, Self::GlobalOpt, Self::Opt>,
-    ) -> Result<Self::ReturnItem> {
-        let client = ctx.state().client();
-        let opt = ctx.opt();
-        let vec_transaction_info = client.chain_get_block_txn_infos2(opt.block_hash)?;
-
-        Ok(vec_transaction_info)
-    }
-}
