--- cmd/starcoin/src/subcommand_vm2/chain/mod.rs
@@ -1,15 +1,0 @@
-// Copyright (c) The Starcoin Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-mod get_events_cmd;
-mod get_txn_cmd;
-mod get_txn_info_cmd;
-mod get_txn_info_list_cmd;
-mod get_txn_infos_cmd;
-pub mod get_txn_proof_cmd;
-
-pub use get_events_cmd::*;
-pub use get_txn_cmd::*;
-pub use get_txn_info_cmd::*;
-pub use get_txn_info_list_cmd::*;
-pub use get_txn_infos_cmd::*;
